refactor(job-store-api): adapterの統一とclientImplの削除
- adapter実装をsrc/adaptersに移動し、コマンドパターンで統一
- src/clientImplディレクトリと関連レガシーコードを削除
- APIルート・テストを新adapterインターフェースに対応
- CommandOutput型を統一し、エラー処理と型安全性を向上
- テストデータの日付をISO8601形式に修正
