+++ apps/job-store-api/src/adapters/error.ts
