+++ apps/job-store-api/test/d1/client.test.ts
@@ -1,9 +1,9 @@
 import { env } from "cloudflare:test";
 import { drizzle } from "drizzle-orm/d1";
 import { expect, it } from "vitest";
-import { createJobStoreResultBuilder } from "../../src/clientImpl";
-import { createJobStoreDBClientAdapter } from "../../src/clientImpl/adapter";
-
+import { createJobStoreDBClientAdapter } from "../../src/adapters";
+import { parse } from "valibot";
+import { insertJobRequestBodySchema } from "@sho/models";
 declare module "cloudflare:test" {
   interface ProvidedEnv {
     DB: D1Database;
@@ -13,9 +13,8 @@ declare module "cloudflare:test" {
 it("求人データを挿入できる", async () => {
   const db = drizzle(env.DB);
   const dbClient = createJobStoreDBClientAdapter(db);
-  const jobStore = createJobStoreResultBuilder(dbClient);
   const jobNumber = "64455-10912";
-  const insertingJob = {
+  const insertingJob = parse(insertJobRequestBodySchema, {
     jobNumber,
     companyName: "Tech Corp",
     jobDescription: "ソフトウェアエンジニアの募集です。",
@@ -25,23 +24,26 @@ it("求人データを挿入できる", async () => {
     employmentType: "正社員",
     workingStartTime: "09:00",
     workingEndTime: "18:00",
-    receivedDate: "2024-06-01",
-    expiryDate: "2024-12-31",
+    receivedDate: "2024-06-01T12:34:56Z",
+    expiryDate: "2024-12-31T23:59:59Z",
     employeeCount: 200,
     occupation: "IT",
     homePage: "https://techcorp.example.com",
     qualifications: " コンピュータサイエンスの学位、3年以上の経験",
-  };
-  const result = await jobStore.insertJob(insertingJob);
-  if (result.isErr()) {
-    console.error(result.error);
+  });
+  const result = await dbClient.execute({
+    type: "InsertJob",
+    payload: insertingJob,
+
+  });
+  expect(result.success).toBe(true);
+  const result2 = await dbClient.execute({
+    type: "FindJobByNumber",
+    jobNumber,
+  });
+  expect(result2.success).toBe(true);
+  if (result2.success) {
+    const returnedJobNumber = result2.job?.jobNumber;
+    expect(returnedJobNumber).toEqual(jobNumber);
   }
-  expect(result.isOk()).toBe(true);
-  const result2 = await jobStore.fetchJob(jobNumber);
-  if (result2.isErr()) {
-    console.error(result2.error);
-  }
-  expect(result2.isOk()).toBe(true);
-  const returnedJobNumber = result2._unsafeUnwrap()?.jobNumber;
-  expect(returnedJobNumber).toEqual(jobNumber);
 });
